// frontend/src/api/models/TaskCompletionResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TaskCompletionResponse
 */
export interface TaskCompletionResponse {
    /**
     * 
     * @type {number}
     * @memberof TaskCompletionResponse
     */
    enrollmentId: number;
    /**
     * 
     * @type {number}
     * @memberof TaskCompletionResponse
     */
    taskId: number;
    /**
     * 
     * @type {Date}
     * @memberof TaskCompletionResponse
     */
    completedAt?: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof TaskCompletionResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {number}
     * @memberof TaskCompletionResponse
     */
    taskCompletionId: number;
}

/**
 * Check if a given object implements the TaskCompletionResponse interface.
 */
export function instanceOfTaskCompletionResponse(value: object): value is TaskCompletionResponse {
    if (!('enrollmentId' in value) || value['enrollmentId'] === undefined) return false;
    if (!('taskId' in value) || value['taskId'] === undefined) return false;
    if (!('taskCompletionId' in value) || value['taskCompletionId'] === undefined) return false;
    return true;
}

export function TaskCompletionResponseFromJSON(json: any): TaskCompletionResponse {
    return TaskCompletionResponseFromJSONTyped(json, false);
}

export function TaskCompletionResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TaskCompletionResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'enrollmentId': json['enrollment_id'],
        'taskId': json['task_id'],
        'completedAt': json['completed_at'] == null ? undefined : (new Date(json['completed_at'])),
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'taskCompletionId': json['task_completion_id'],
    };
}

export function TaskCompletionResponseToJSON(json: any): TaskCompletionResponse {
    return TaskCompletionResponseToJSONTyped(json, false);
}

export function TaskCompletionResponseToJSONTyped(value?: TaskCompletionResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'enrollment_id': value['enrollmentId'],
        'task_id': value['taskId'],
        'completed_at': value['completedAt'] == null ? undefined : ((value['completedAt'] as any).toISOString()),
        'is_active': value['isActive'],
        'task_completion_id': value['taskCompletionId'],
    };
}