// frontend/src/api/models/TaskCompletionCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TaskCompletionCreate
 */
export interface TaskCompletionCreate {
    /**
     * 
     * @type {number}
     * @memberof TaskCompletionCreate
     */
    enrollmentId: number;
    /**
     * 
     * @type {number}
     * @memberof TaskCompletionCreate
     */
    taskId: number;
    /**
     * 
     * @type {Date}
     * @memberof TaskCompletionCreate
     */
    completedAt?: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof TaskCompletionCreate
     */
    isActive?: boolean;
}

/**
 * Check if a given object implements the TaskCompletionCreate interface.
 */
export function instanceOfTaskCompletionCreate(value: object): value is TaskCompletionCreate {
    if (!('enrollmentId' in value) || value['enrollmentId'] === undefined) return false;
    if (!('taskId' in value) || value['taskId'] === undefined) return false;
    return true;
}

export function TaskCompletionCreateFromJSON(json: any): TaskCompletionCreate {
    return TaskCompletionCreateFromJSONTyped(json, false);
}

export function TaskCompletionCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TaskCompletionCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'enrollmentId': json['enrollment_id'],
        'taskId': json['task_id'],
        'completedAt': json['completed_at'] == null ? undefined : (new Date(json['completed_at'])),
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
    };
}

export function TaskCompletionCreateToJSON(json: any): TaskCompletionCreate {
    return TaskCompletionCreateToJSONTyped(json, false);
}

export function TaskCompletionCreateToJSONTyped(value?: TaskCompletionCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'enrollment_id': value['enrollmentId'],
        'task_id': value['taskId'],
        'completed_at': value['completedAt'] == null ? undefined : ((value['completedAt'] as any).toISOString()),
        'is_active': value['isActive'],
    };
}